requirements.txt
pandas
requests
snowflake-connector-python
pyarrow